requirements.txt
bagit==1.5.4
boto3==1.2.2
botocore==1.3.9
docutils==0.12
futures==2.2.0
jmespath==0.9.0
python-dateutil==2.4.2
PyYAML==3.11
six==1.10.0
wheel==0.24.0
